GenSwarm/requirements.txt
dashscope
openai==1.8.0
pydantic==2.5.3
PyYAML==6.0.1
typer==0.9.0
loguru==0.5.3
tenacity==8.2.3
numpy==1.26.3
matplotlib==3.8.0
scipy~=1.12.0
opencv-python==4.9.0.80
pynput==1.7.6
anthropic
pylint
coverage
imageio
pre-commit
rospkg
radon
tabulate
pygame~=2.5.2
gymnasium~=0.29.1
pillow~=10.2.0
box2d-py
fastdtw==0.3.4
paho-mqtt==1.6.1
rospy_message_converter
pygments
rich
